Create the main audit code
This creates the main auditing code, including a class for representing
audit failures, one representing a set of audit checks, and three
separate checks to perform on a pull request.  The audit checks
implemented so far are the "self_merge" check; the "merged_before_tests"
check, which complains if tests were not finished before the pull
request was merged; and the "tests_failed" check, which complains if
the pull request was merged with the tests failing.  The audit checks
involving tests make their determinations based on the status of the
last commit in the pull request.