Obtain Github handles
This brings in a simple caching routine to obtain handles for the
Github API.  The routine added allows access to any arbitrary
Github API URL with any arbitrary set of user credentials, even
prompting for any passwords needed.  The routine also caches
handles, allowing them to be reused later.  This design will allow
any arbitrary set of repositories to be audited in a single run.